relocated and refactored utils classes
