+++ openvalidation-antlr/src/main/java/io/openvalidation/antlr/transformation/parsetree/PTRuleConstrainedTransformer.java
@@ -18,7 +18,6 @@ package io.openvalidation.antlr.transformation.parsetree;
 
 import io.openvalidation.antlr.generated.mainLexer;
 import io.openvalidation.antlr.generated.mainParser;
-import io.openvalidation.antlr.transformation.ParseTreeUtils;
 import io.openvalidation.antlr.transformation.TransformerBase;
 import io.openvalidation.antlr.transformation.TransformerContext;
 import io.openvalidation.common.ast.ASTItem;
